Base completa;

+++ Models/Produto.cs
@@ -8,7 +8,7 @@ namespace CoffeeShop.Models
         public Guid Id { get; set; } = Guid.NewGuid();
         public string Nome { get; set; }
         public string Descricao { get; set; }
-        public decimal Preco { get; set; }
+        public int Preco { get; set; }
         public int QuantidadeEstoque { get; set; }
         public Guid FornecedorId { get; set; }
         public Fornecedor ? Fornecedor { get; set; }
